+++ src/main/java/ch/bzz/pokemon/service/TypService.java
@@ -1,6 +1,7 @@
 package ch.bzz.pokemon.service;
 
 import ch.bzz.pokemon.data.DataHandler;
+import ch.bzz.pokemon.model.Pokemon;
 import ch.bzz.pokemon.model.Typ;
 import com.sun.xml.internal.bind.v2.model.core.ID;
 
@@ -19,16 +20,24 @@ import java.util.UUID;
 public class TypService {
     /**
      * reads a list of all types
-     * @return
+     * @return a list of all types
      */
     @GET
     @Path("list")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response listTypes(){
-        List<Typ> typList = DataHandler.readAllTypes();
+    public Response listTypes(
+            @CookieParam("userRole") String userRole
+    ){
+        int httpsStatus = 0;
+        if (userRole == null || userRole.equals("guest")){
+            httpsStatus = 403;
+        }else if (userRole.equals("admin") || userRole.equals("user")) {
+            httpsStatus = 200;
+            List<Typ> typList = DataHandler.readAllTypes();
+        }
         return Response
                 .status(200)
-                .entity(typList)
+                .entity(httpsStatus)
                 .build();
     }
     /**
@@ -42,51 +51,63 @@ public class TypService {
     public Response readTyp(
             @NotEmpty
             @Pattern(regexp="ID-\\d{1,3}")
-            @QueryParam("uuid") String typUUID
+            @QueryParam("uuid") String typUUID,
+            @CookieParam("userRole") String userRole
     ){
-        Typ typ = DataHandler.readTypByUUID(typUUID);
-        int httpsStatus;
-        if (typ == null){
-            httpsStatus = 404;
-        }else {
-            httpsStatus = 288;
+        int httpsStatus = 0;
+        if (userRole == null || userRole.equals("guest")){
+            httpsStatus = 403;
+        }else if (userRole.equals("admin") || userRole.equals("user"))  {
+            httpsStatus = 200;
+            Typ typ = DataHandler.readTypByUUID(typUUID);
         }
         return Response
-                .status(httpsStatus)
-                .entity(typ)
+                .status(200)
+                .entity(httpsStatus)
                 .build();
     }
 
     /**
-     *
-     * @return Response
+     * create a typ by UUID
+     * @return response
      */
     @POST
     @Path("create")
     @Produces(MediaType.TEXT_PLAIN)
     public Response insertTyp(
-            @Valid @BeanParam Typ typ
+            @Valid @BeanParam Typ typ,
+            @CookieParam("userRole") String userRole
     ){
-        typ.setTypUUID(UUID.randomUUID().toString());
-        typ.setTyp(typ.getTypUUID());
-        DataHandler.insertTyp(typ);
+        int httpsStatus;
+        if (userRole == null || !userRole.equals("admin")){
+            httpsStatus = 403;
+        }else {
+            httpsStatus = 200;
+            typ.setTypUUID(UUID.randomUUID().toString());
+            typ.setTyp(typ.getTypUUID());
+            DataHandler.insertTyp(typ);
+        }
         return Response
                 .status(200)
-                .entity("")
+                .entity(httpsStatus)
                 .build();
     }
 
     /**
-     *
+     * updates a typ by UUID
      * @return Response
      */
     @POST
     @Path("update")
     @Produces(MediaType.TEXT_PLAIN)
     public Response updateTyp(
-            @Valid @BeanParam Typ typ
+            @Valid @BeanParam Typ typ,
+            @CookieParam("userRole") String userRole
 
     ){
+        if (userRole == null || !userRole.equals("admin")) {
+            return Response.status(403).entity("").build();
+        }
         int httpStatus = 200;
         Typ oldTyp = DataHandler.readTypByUUID(typ.getTypUUID());
         if (oldTyp != null){
@@ -102,7 +123,7 @@ public class TypService {
     }
 
     /**
-     * delets a pokemon indentified by its uuid
+     * delets a typ indentified by its uuid
      * @param typUUID the key
      * @return Response
      */
@@ -110,8 +131,12 @@ public class TypService {
     @Path("delete")
     @Produces(MediaType.TEXT_PLAIN)
     public Response deleteTyp(
-            @QueryParam("uuid") String typUUID
+            @QueryParam("uuid") String typUUID,
+                @CookieParam("userRole") String userRole
     ){
+        if (userRole == null || !userRole.equals("admin")) {
+            return Response.status(403).entity("").build();
+        }
         int httpStatus = 200;
         if (!DataHandler.deleteTyp(typUUID)){
             httpStatus = 410;
